refactor(payment): CHECKOUT-3011 Move Paypal related strategies to Paypal folder
