+++ src/payment/strategies/paypal/index.ts
@@ -1,3 +1,5 @@
 export * from './paypal-sdk';
 
 export { default as PaypalScriptLoader } from './paypal-script-loader';
+export { default as PaypalExpressPaymentStrategy } from './paypal-express-payment-strategy';
+export { default as PaypalProPaymentStrategy } from './paypal-pro-payment-strategy';
